refactor transformer support code into modules
