+++ crates/bitnet-transformer/src/kv_cache.rs
@@ -1,0 +1,125 @@
+//! Key/value cache storage for transformer decoding.
+
+use bitnet_common::{BitNetConfig, BitNetError, Result};
+use candle_core::{DType, Device, Tensor};
+
+/// KV cache for a single transformer layer.
+pub struct LayerKVCache {
+    pub k: Tensor,
+    pub v: Tensor,
+    pub seq_len: usize,
+    pub max_seq_len: usize,
+    pub n_kv_heads: usize, // Store the number of KV heads for validation
+}
+
+impl LayerKVCache {
+    pub fn new(
+        batch_size: usize,
+        n_kv_heads: usize, // Changed from n_heads to n_kv_heads
+        max_seq_len: usize,
+        head_dim: usize,
+        device: &Device,
+    ) -> Result<Self> {
+        let k =
+            Tensor::zeros(&[batch_size, n_kv_heads, max_seq_len, head_dim], DType::F32, device)?;
+        let v =
+            Tensor::zeros(&[batch_size, n_kv_heads, max_seq_len, head_dim], DType::F32, device)?;
+
+        Ok(Self { k, v, seq_len: 0, max_seq_len, n_kv_heads })
+    }
+
+    /// Append new K/V tensors to the cache
+    ///
+    /// **Performance note**: The clones on first append (lines 1130-1131) are necessary
+    /// because we accept `&Tensor` but need to store owned tensors. Candle's `Tensor::clone()`
+    /// is cheap - it only increments the Arc reference count, not a deep data copy.
+    /// Subsequent appends use `Tensor::cat` which allocates new tensors regardless.
+    ///
+    /// To eliminate these clones would require API changes to accept owned tensors,
+    /// which would complicate calling code.
+    pub fn append(&mut self, k_new: &Tensor, v_new: &Tensor) -> Result<()> {
+        // Expect shapes: k: [B,HKV,T_new,Hd], v: [B,HKV,T_new,Hd] where HKV = n_kv_heads
+        let new_seq_len = k_new.dims()[2];
+
+        // Validate that the incoming tensors have the expected number of KV heads
+        let k_heads = k_new.dims()[1];
+        if k_heads != self.n_kv_heads {
+            return Err(BitNetError::Validation(format!(
+                "KV cache expects {} heads, but received K tensor with {} heads",
+                self.n_kv_heads, k_heads
+            )));
+        }
+
+        if self.seq_len == 0 {
+            // First append: clone is necessary (Arc increment only, not deep copy)
+            self.k = k_new.clone();
+            self.v = v_new.clone();
+        } else {
+            // Concatenate along time dimension (dim=2)
+            if self.seq_len + new_seq_len > self.max_seq_len {
+                return Err(BitNetError::from(candle_core::Error::Msg(
+                    "KV cache overflow".to_string(),
+                )));
+            }
+            // Tensor::cat allocates new tensor - no optimization possible here
+            self.k = Tensor::cat(&[&self.k, k_new], 2)?;
+            self.v = Tensor::cat(&[&self.v, v_new], 2)?;
+        }
+
+        self.seq_len += new_seq_len;
+        Ok(())
+    }
+
+    pub fn clear(&mut self) {
+        self.seq_len = 0;
+    }
+}
+
+/// Full KV Cache for all layers
+pub struct KVCache {
+    pub layers: Vec<LayerKVCache>,
+}
+
+impl KVCache {
+    pub fn new(config: &BitNetConfig, batch_size: usize, device: &Device) -> Result<Self> {
+        let n_layers = config.model.num_layers;
+        let n_heads = config.model.num_heads;
+        let hidden_size = config.model.hidden_size;
+
+        // Validate shape assumptions before calculating dimensions
+        if config.model.attention_head_dim.is_none() && !hidden_size.is_multiple_of(n_heads) {
+            return Err(BitNetError::Validation(format!(
+                "KVCache: hidden_size {} not divisible by num_heads {}",
+                hidden_size, n_heads
+            )));
+        }
+
+        let n_kv_heads = config.model.num_key_value_heads.max(1).min(n_heads);
+        if !n_heads.is_multiple_of(n_kv_heads) {
+            return Err(BitNetError::Validation(format!(
+                "KVCache: num_heads {} not divisible by num_key_value_heads {}",
+                n_heads, n_kv_heads
+            )));
+        }
+
+        let head_dim = config.model.attention_head_dim.unwrap_or_else(|| hidden_size / n_heads);
+        let max_seq_len = config.model.max_position_embeddings;
+
+        let mut layers = Vec::with_capacity(n_layers);
+        for _ in 0..n_layers {
+            layers.push(LayerKVCache::new(batch_size, n_kv_heads, max_seq_len, head_dim, device)?);
+        }
+
+        Ok(Self { layers })
+    }
+
+    pub fn layer_mut(&mut self, idx: usize) -> Option<&mut LayerKVCache> {
+        self.layers.get_mut(idx)
+    }
+
+    pub fn clear(&mut self) {
+        for layer in &mut self.layers {
+            layer.clear();
+        }
+    }
+}
