+++ crates/bitnet-transformer/src/lib.rs
@@ -7,251 +7,18 @@ use bitnet_qk256_dispatch::{
     strict_cuda_bitnet_backend_requested,
 };
 use bitnet_rope::{build_tables as build_rope_tables, resolve_base as resolve_rope_base};
-use candle_core::{DType, Device, Module, Tensor};
+use candle_core::{Device, Module, Tensor};
 use candle_nn::{LayerNorm, Linear, VarBuilder};
-use std::sync::OnceLock;
-
 mod attention_forward;
+mod diagnostics;
+mod kv_cache;
 
-#[derive(Clone, Debug)]
-struct QwenTraceConfig {
-    path: Option<std::path::PathBuf>,
-    stderr_enabled: bool,
-    layer: usize,
-}
-
-fn env_flag(name: &str) -> bool {
-    std::env::var(name).is_ok()
-}
-
-fn env_flag_eq_1(name: &str) -> bool {
-    std::env::var(name).as_deref() == Ok("1")
-}
-
-fn qwen_trace_config() -> &'static QwenTraceConfig {
-    static CONFIG: OnceLock<QwenTraceConfig> = OnceLock::new();
-    CONFIG.get_or_init(|| QwenTraceConfig {
-        path: std::env::var("BITNET_QWEN_TRACE_JSONL")
-            .ok()
-            .filter(|value| !value.trim().is_empty())
-            .map(std::path::PathBuf::from),
-        stderr_enabled: env_flag_eq_1("BITNET_QWEN_TRACE"),
-        layer: std::env::var("BITNET_QWEN_TRACE_LAYER")
-            .ok()
-            .and_then(|value| value.parse::<usize>().ok())
-            .unwrap_or(0),
-    })
-}
-
-fn debug_attn_enabled() -> bool {
-    static ENABLED: OnceLock<bool> = OnceLock::new();
-    *ENABLED.get_or_init(|| env_flag("DEBUG_ATTN"))
-}
-
-fn debug_attn_scale_enabled() -> bool {
-    static ENABLED: OnceLock<bool> = OnceLock::new();
-    *ENABLED.get_or_init(|| env_flag("BITNET_DEBUG_ATTN_SCALE"))
-}
-
-fn debug_gqa_enabled() -> bool {
-    static ENABLED: OnceLock<bool> = OnceLock::new();
-    *ENABLED.get_or_init(|| env_flag("BITNET_DEBUG_GQA"))
-}
-
-fn debug_rope_enabled() -> bool {
-    static ENABLED: OnceLock<bool> = OnceLock::new();
-    *ENABLED.get_or_init(|| env_flag("BITNET_DEBUG_ROPE"))
-}
-
-fn debug_rmsnorm_enabled() -> bool {
-    static ENABLED: OnceLock<bool> = OnceLock::new();
-    *ENABLED.get_or_init(|| env_flag("BITNET_DEBUG_RMSNORM"))
-}
-
-fn debug_mlp_enabled() -> bool {
-    static ENABLED: OnceLock<bool> = OnceLock::new();
-    *ENABLED.get_or_init(|| env_flag("BITNET_DEBUG_MLP"))
-}
-
-fn trace_rms_enabled() -> bool {
-    static ENABLED: OnceLock<bool> = OnceLock::new();
-    *ENABLED.get_or_init(|| env_flag_eq_1("BITNET_TRACE_RMS"))
-}
-
-fn qwen_trace_path() -> Option<std::path::PathBuf> {
-    qwen_trace_config().path.clone()
-}
-
-fn qwen_trace_enabled() -> bool {
-    let config = qwen_trace_config();
-    config.path.is_some() || config.stderr_enabled
-}
-
-fn qwen_trace_active() -> bool {
-    qwen_trace_enabled() && std::env::var("BITNET_QWEN_TRACE_ACTIVE").as_deref() == Ok("1")
-}
-
-fn qwen_trace_layer_enabled(layer_idx: usize) -> bool {
-    if !qwen_trace_active() {
-        return false;
-    }
-    qwen_trace_config().layer == layer_idx
-}
-
-fn qwen_trace_escape(value: &str) -> String {
-    value.replace('\\', "\\\\").replace('"', "\\\"").replace('\n', "\\n").replace('\r', "\\r")
-}
-
-fn qwen_trace_number(value: f64) -> String {
-    if value.is_finite() { format!("{value:.9}") } else { "null".to_string() }
-}
-
-fn qwen_trace_write_line(line: &str) {
-    if let Some(path) = qwen_trace_path() {
-        if let Some(parent) = path.parent()
-            && let Err(err) = std::fs::create_dir_all(parent)
-        {
-            eprintln!("qwen_trace_write_failed: create_dir_all {}: {err}", parent.display());
-            return;
-        }
-        match std::fs::OpenOptions::new().create(true).append(true).open(&path) {
-            Ok(mut file) => {
-                if let Err(err) = std::io::Write::write_all(&mut file, line.as_bytes())
-                    .and_then(|_| std::io::Write::write_all(&mut file, b"\n"))
-                {
-                    eprintln!("qwen_trace_write_failed: {}: {err}", path.display());
-                }
-            }
-            Err(err) => eprintln!("qwen_trace_write_failed: {}: {err}", path.display()),
-        }
-    } else if std::env::var("BITNET_QWEN_TRACE").as_deref() == Ok("1") {
-        eprintln!("{line}");
-    }
-}
-
-fn qwen_trace_event(stage: &str, fields_json: &str) {
-    if !qwen_trace_enabled() {
-        return;
-    }
-    let step = std::env::var("BITNET_QWEN_TRACE_STEP").unwrap_or_else(|_| "null".to_string());
-    qwen_trace_write_line(&format!(
-        "{{\"kind\":\"qwen_trace_event\",\"stage\":\"{}\",\"step\":{},{} }}",
-        qwen_trace_escape(stage),
-        step,
-        fields_json
-    ));
-}
-
-fn qwen_trace_tensor(
-    stage: &str,
-    layer_idx: Option<usize>,
-    tensor: &Tensor,
-) -> candle_core::Result<()> {
-    if !qwen_trace_active() {
-        return Ok(());
-    }
-    if let Some(layer_idx) = layer_idx
-        && !qwen_trace_layer_enabled(layer_idx)
-    {
-        return Ok(());
-    }
-
-    let tensor_f32 =
-        if tensor.dtype() == DType::F32 { tensor.clone() } else { tensor.to_dtype(DType::F32)? };
-    let values = tensor_f32.flatten_all()?.to_vec1::<f32>()?;
-    let mut finite_count = 0usize;
-    let mut nonfinite_count = 0usize;
-    let mut sum = 0.0f64;
-    let mut sum_sq = 0.0f64;
-    let mut min = f64::INFINITY;
-    let mut max = f64::NEG_INFINITY;
-    let mut checksum = 0.0f64;
-    for (idx, value) in values.iter().enumerate() {
-        let value = *value as f64;
-        if value.is_finite() {
-            finite_count += 1;
-            sum += value;
-            sum_sq += value * value;
-            min = min.min(value);
-            max = max.max(value);
-            if idx < 4096 {
-                checksum += value * ((idx % 257) + 1) as f64;
-            }
-        } else {
-            nonfinite_count += 1;
-        }
-    }
-
-    let denom = finite_count.max(1) as f64;
-    let mean = sum / denom;
-    let rms = (sum_sq / denom).sqrt();
-    let sample = values
-        .iter()
-        .take(8)
-        .map(|value| qwen_trace_number(*value as f64))
-        .collect::<Vec<_>>()
-        .join(",");
-    let dims = tensor.dims().iter().map(|dim| dim.to_string()).collect::<Vec<_>>().join(",");
-    let layer_json = layer_idx.map(|idx| idx.to_string()).unwrap_or_else(|| "null".to_string());
-    let step = std::env::var("BITNET_QWEN_TRACE_STEP").unwrap_or_else(|_| "null".to_string());
-
-    qwen_trace_write_line(&format!(
-        "{{\"kind\":\"qwen_trace_tensor\",\"stage\":\"{}\",\"step\":{},\"layer\":{},\"dtype\":\"{:?}\",\"dims\":[{}],\"len\":{},\"finite\":{},\"nonfinite\":{},\"mean\":{},\"rms\":{},\"min\":{},\"max\":{},\"checksum\":{},\"sample\":[{}]}}",
-        qwen_trace_escape(stage),
-        step,
-        layer_json,
-        tensor.dtype(),
-        dims,
-        values.len(),
-        finite_count,
-        nonfinite_count,
-        qwen_trace_number(mean),
-        qwen_trace_number(rms),
-        qwen_trace_number(min),
-        qwen_trace_number(max),
-        qwen_trace_number(checksum),
-        sample
-    ));
-    Ok(())
-}
-
-/// Debug helper for tensor statistics (only runs if DEBUG_ATTN env var is set)
-fn dbg_stats(tag: &str, t: &Tensor) -> candle_core::Result<()> {
-    if debug_attn_enabled() {
-        let mean = t.mean_all()?.to_scalar::<f32>()?;
-        // Compute std manually: sqrt(E[(x - mean)^2])
-        let diff = t.broadcast_sub(&t.mean_all()?)?;
-        let variance = diff.sqr()?.mean_all()?;
-        let std = variance.sqrt()?.to_scalar::<f32>()?;
-        eprintln!("[dbg] {tag}: mean={mean:.6} std={std:.6}");
-    }
-    Ok(())
-}
-
-/// Debug helper for checking finite values
-fn dbg_finite(tag: &str, t: &Tensor) -> candle_core::Result<()> {
-    if debug_attn_enabled() {
-        let v: Vec<f32> = t.flatten_all()?.to_vec1()?;
-        let n = v.len().min(4096);
-        let mut n_nan = 0;
-        let mut n_inf = 0;
-        for &x in &v[..n] {
-            if !x.is_finite() {
-                if x.is_nan() {
-                    n_nan += 1;
-                } else {
-                    n_inf += 1;
-                }
-            }
-        }
-        if n_nan + n_inf > 0 {
-            eprintln!(
-                "⚠️  [dbg] {tag}: non-finite values: NaN={n_nan} Inf={n_inf} (in first {n} elems)"
-            );
-        }
-    }
-    Ok(())
-}
+use diagnostics::{
+    dbg_finite, dbg_stats, debug_attn_enabled, debug_attn_scale_enabled, debug_gqa_enabled,
+    debug_mlp_enabled, debug_rmsnorm_enabled, debug_rope_enabled, qwen_trace_event,
+    qwen_trace_layer_enabled, qwen_trace_number, qwen_trace_tensor, trace_rms_enabled,
+};
+pub use kv_cache::{KVCache, LayerKVCache};
 
 /// Helper to create linear layers with optional bias tensors.
 fn linear_with_optional_bias(
@@ -1030,126 +797,6 @@ fn eps_from_config(config: &BitNetConfig) -> f64 {
 }
 
 /// KV Cache for a single layer
-pub struct LayerKVCache {
-    pub k: Tensor,
-    pub v: Tensor,
-    pub seq_len: usize,
-    pub max_seq_len: usize,
-    pub n_kv_heads: usize, // Store the number of KV heads for validation
-}
-
-impl LayerKVCache {
-    pub fn new(
-        batch_size: usize,
-        n_kv_heads: usize, // Changed from n_heads to n_kv_heads
-        max_seq_len: usize,
-        head_dim: usize,
-        device: &Device,
-    ) -> Result<Self> {
-        let k =
-            Tensor::zeros(&[batch_size, n_kv_heads, max_seq_len, head_dim], DType::F32, device)?;
-        let v =
-            Tensor::zeros(&[batch_size, n_kv_heads, max_seq_len, head_dim], DType::F32, device)?;
-
-        Ok(Self { k, v, seq_len: 0, max_seq_len, n_kv_heads })
-    }
-
-    /// Append new K/V tensors to the cache
-    ///
-    /// **Performance note**: The clones on first append (lines 1130-1131) are necessary
-    /// because we accept `&Tensor` but need to store owned tensors. Candle's `Tensor::clone()`
-    /// is cheap - it only increments the Arc reference count, not a deep data copy.
-    /// Subsequent appends use `Tensor::cat` which allocates new tensors regardless.
-    ///
-    /// To eliminate these clones would require API changes to accept owned tensors,
-    /// which would complicate calling code.
-    pub fn append(&mut self, k_new: &Tensor, v_new: &Tensor) -> Result<()> {
-        // Expect shapes: k: [B,HKV,T_new,Hd], v: [B,HKV,T_new,Hd] where HKV = n_kv_heads
-        let new_seq_len = k_new.dims()[2];
-
-        // Validate that the incoming tensors have the expected number of KV heads
-        let k_heads = k_new.dims()[1];
-        if k_heads != self.n_kv_heads {
-            return Err(BitNetError::Validation(format!(
-                "KV cache expects {} heads, but received K tensor with {} heads",
-                self.n_kv_heads, k_heads
-            )));
-        }
-
-        if self.seq_len == 0 {
-            // First append: clone is necessary (Arc increment only, not deep copy)
-            self.k = k_new.clone();
-            self.v = v_new.clone();
-        } else {
-            // Concatenate along time dimension (dim=2)
-            if self.seq_len + new_seq_len > self.max_seq_len {
-                return Err(BitNetError::from(candle_core::Error::Msg(
-                    "KV cache overflow".to_string(),
-                )));
-            }
-            // Tensor::cat allocates new tensor - no optimization possible here
-            self.k = Tensor::cat(&[&self.k, k_new], 2)?;
-            self.v = Tensor::cat(&[&self.v, v_new], 2)?;
-        }
-
-        self.seq_len += new_seq_len;
-        Ok(())
-    }
-
-    pub fn clear(&mut self) {
-        self.seq_len = 0;
-    }
-}
-
-/// Full KV Cache for all layers
-pub struct KVCache {
-    pub layers: Vec<LayerKVCache>,
-}
-
-impl KVCache {
-    pub fn new(config: &BitNetConfig, batch_size: usize, device: &Device) -> Result<Self> {
-        let n_layers = config.model.num_layers;
-        let n_heads = config.model.num_heads;
-        let hidden_size = config.model.hidden_size;
-
-        // Validate shape assumptions before calculating dimensions
-        if config.model.attention_head_dim.is_none() && !hidden_size.is_multiple_of(n_heads) {
-            return Err(BitNetError::Validation(format!(
-                "KVCache: hidden_size {} not divisible by num_heads {}",
-                hidden_size, n_heads
-            )));
-        }
-
-        let n_kv_heads = config.model.num_key_value_heads.max(1).min(n_heads);
-        if !n_heads.is_multiple_of(n_kv_heads) {
-            return Err(BitNetError::Validation(format!(
-                "KVCache: num_heads {} not divisible by num_key_value_heads {}",
-                n_heads, n_kv_heads
-            )));
-        }
-
-        let head_dim = config.model.attention_head_dim.unwrap_or_else(|| hidden_size / n_heads);
-        let max_seq_len = config.model.max_position_embeddings;
-
-        let mut layers = Vec::with_capacity(n_layers);
-        for _ in 0..n_layers {
-            layers.push(LayerKVCache::new(batch_size, n_kv_heads, max_seq_len, head_dim, device)?);
-        }
-
-        Ok(Self { layers })
-    }
-
-    pub fn layer_mut(&mut self, idx: usize) -> Option<&mut LayerKVCache> {
-        self.layers.get_mut(idx)
-    }
-
-    pub fn clear(&mut self) {
-        for layer in &mut self.layers {
-            layer.clear();
-        }
-    }
-}
-
 /// Complete Transformer Model
 pub struct TransformerModel {
     pub config: BitNetConfig,
@@ -1769,6 +1416,7 @@ impl TransformerModel {
 mod tests {
     use super::*;
     use bitnet_common::config::ModelConfig;
+    use candle_core::DType;
     use candle_nn::RmsNorm;
     use serial_test::serial;
     use std::collections::HashMap;
